requirements-dev.txt
mypy
pandas-stubs
types-tqdm
types-requests
types-regex